api/requirements.txt
amqp==2.4.2
aniso8601==6.0.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
billiard==3.5.0.5
celery==4.2.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coloredlogs==10.0
dill==0.2.9
dnspython==1.16.0
docopt==0.6.2
Flask==1.0.2
Flask-Cors==3.0.7
flask-pytest==0.0.5
Flask-Redis==0.3.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
flask-swagger==0.2.13
flask-swagger-ui==3.20.9
flower==0.9.3
humanfriendly==4.18
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10
kafka==1.3.5
kafka-python==1.4.6
kombu==4.4.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==6.0.0
netaddr==0.7.19
pluggy==0.9.0
py==1.8.0
pylint==2.3.1
pytest==4.4.1
python-dotenv==0.10.3
python-etcd==0.4.5
pytz==2018.9
PyYAML==5.1.1
redis==3.2.1
requests==2.21.0
six==1.12.0
tornado==5.1.1
typed-ast==1.4.0
urllib3==1.24.1
vine==1.3.0
Werkzeug==0.15.0
wrapt==1.11.2
